Apply changes to files in /second_q/ and undo changes in /algorithms/.
