+++ test/algorithms/initial_points/resources/test_mp2_initial_point.json
