+++ test/second_q/algorithms/initial_points/test_mp2_initial_point.py
@@ -58,10 +58,10 @@ class TestMP2InitialPoint(QiskitNatureTestCase):
         self.mock_ansatz.excitation_list = self.excitation_list
 
         electronic_energy = Mock(spec=ElectronicEnergy)
-        electronic_energy.orbital_energies = Mock(spec=np.ndarray)
+        electronic_energy.orbital_energies = np.asarray([0])
         electronic_energy.reference_energy = 123.45
         electronic_integrals = Mock(spec=ElectronicIntegrals)
-        electronic_integrals.get_matrix = Mock(return_value=[0])
+        electronic_integrals.get_matrix = Mock(return_value=np.asarray([0]))
         electronic_energy.get_electronic_integral = Mock(return_value=electronic_integrals)
         self.mock_grouped_property = Mock(spec=GroupedSecondQuantizedProperty)
         self.mock_grouped_property.get_property = Mock(return_value=electronic_energy)
@@ -130,14 +130,12 @@ class TestMP2InitialPoint(QiskitNatureTestCase):
         electronic_energy = Mock(spec=ElectronicEnergy)
         electronic_energy.get_electronic_integral = Mock(return_value=electronic_integrals)
         electronic_energy.orbital_energies = None
-        self.mock_grouped_property.get_property = Mock(return_value=electronic_energy)
-
-        ansatz = Mock(spec=UCC)
-        ansatz.excitation_list = self.excitation_list
+        grouped_property = Mock(spec=GroupedSecondQuantizedProperty)
+        grouped_property.get_property = Mock(return_value=electronic_energy)
 
         mp2_initial_point = MP2InitialPoint()
         with self.assertRaises(QiskitNatureError):
-            mp2_initial_point.compute(ansatz=ansatz, grouped_property=self.mock_grouped_property)
+            mp2_initial_point.compute(ansatz=self.mock_ansatz, grouped_property=grouped_property)
 
     def test_set_excitations_directly(self):
         """Test when setting excitations directly."""
@@ -190,78 +188,74 @@ class TestMP2InitialPoint(QiskitNatureTestCase):
         grouped_property = Mock(spec=GroupedSecondQuantizedProperty)
         grouped_property.get_property = Mock(return_value=electronic_energy)
 
-        ansatz = Mock(spec=UCC)
-        ansatz.excitation_list = self.excitation_list
-
         mp2_initial_point = MP2InitialPoint()
         with self.assertRaises(NotImplementedError):
-            mp2_initial_point.compute(ansatz=ansatz, grouped_property=grouped_property)
+            mp2_initial_point.compute(ansatz=self.mock_ansatz, grouped_property=grouped_property)
 
-    @file_data("./resources/test_mp2_initial_point.json")
+    @unittest.skipIf(not optionals.HAS_PYSCF, "pyscf not available.")
+    @data("H 0 0 0; H 0 0 0.7", "Li 0 0 0; H 0 0 1.6")
     def test_mp2_initial_point_with_real_molecules(
         self,
-        atom1,
-        atom2,
-        distance,
-        coefficients,
-        energy_correction,
-        energy_corrections,
-        energy,
-        excitations,
+        atom,
     ):
-        """Test MP2InitialPoint with real molecules.
+        """Test MP2InitialPoint with real molecules."""
+        from pyscf import gto  # pylint: disable=import-error
 
-        Full excitation sequences generated using:
+        # Compute the PySCF result
+        pyscf_mol = gto.M(atom=atom, basis="sto3g", verbose=0)
+        pyscf_mp = pyscf_mol.MP2().run(verbose=0)
 
-        .. code-block:: python
+        driver = PySCFDriver(atom=atom, basis="sto3g")
 
-            converter = QubitConverter(JordanWignerMapper()
-            ansatz = UCCSD(
-                qubit_converter=converter,
-                num_particles=num_particles,
-                num_spin_orbitals=num_spin_orbitals,
-            )
-            _ = ansatz.operators
-            excitations = ansatz.excitation_list
-        """
+        problem = ElectronicStructureProblem(driver)
+        problem.second_q_ops()
+        grouped_property = problem.grouped_property_transformed
+        particle_number = grouped_property.get_property("ParticleNumber")
 
-        molecule = Molecule(geometry=[[atom1, [0.0, 0.0, 0.0]], [atom2, [0.0, 0.0, distance]]])
+        num_particles = (particle_number.num_alpha, particle_number.num_beta)
+        num_spin_orbitals = particle_number.num_spin_orbitals
 
-        try:
-            driver = ElectronicStructureMoleculeDriver(
-                molecule, basis="sto3g", driver_type=ElectronicStructureDriverType.PYSCF
-            )
-            problem = ElectronicStructureProblem(driver)
-            problem.second_q_ops()
-        except MissingOptionalLibraryError:
-            self.skipTest("PySCF driver does not appear to be installed.")
+        qubit_converter = QubitConverter(mapper=JordanWignerMapper(), two_qubit_reduction=True)
 
-        driver_result = problem.grouped_property_transformed
-
-        ansatz = Mock(spec=UCC)
-        ansatz.excitation_list = excitations
+        initial_state = HartreeFock(
+            num_spin_orbitals=num_spin_orbitals,
+            num_particles=num_particles,
+            qubit_converter=qubit_converter,
+        )
+        ansatz = UCC(
+            num_spin_orbitals=num_spin_orbitals,
+            num_particles=num_particles,
+            excitations="sd",
+            qubit_converter=qubit_converter,
+            initial_state=initial_state,
+        )
 
         mp2_initial_point = MP2InitialPoint()
-        mp2_initial_point.grouped_property = driver_result
+        mp2_initial_point.grouped_property = grouped_property
         mp2_initial_point.ansatz = ansatz
 
-        with self.subTest("Test MP2 initial point array."):
-            np.testing.assert_array_almost_equal(
-                mp2_initial_point.to_numpy_array(), coefficients, decimal=6
-            )
-
-        with self.subTest("Test MP2 energy corrections."):
-            np.testing.assert_array_almost_equal(
-                mp2_initial_point.get_energy_corrections(), energy_corrections, decimal=6
-            )
-
         with self.subTest("Test overall MP2 energy correction."):
-            np.testing.assert_array_almost_equal(
-                mp2_initial_point.get_energy_correction(), energy_correction, decimal=6
+            np.testing.assert_almost_equal(
+                mp2_initial_point.get_energy_correction(), pyscf_mp.e_corr, decimal=10
             )
 
         with self.subTest("Test absolute MP2 energy."):
-            np.testing.assert_array_almost_equal(mp2_initial_point.get_energy(), energy, decimal=6)
+            np.testing.assert_almost_equal(
+                mp2_initial_point.get_energy(), pyscf_mp.e_tot, decimal=10
+            )
+
+        with self.subTest("Test absolute MP2 energy."):
+            print("pyscf", pyscf_mp.t2.shape)
+            print(pyscf_mp.t2)
+            print(pyscf_mp.t2.shape)
+            print("qiskit", mp2_initial_point._t2.shape)
+            diff = pyscf_mp.t2 - mp2_initial_point._t2
+            diff[np.abs(diff) < 1e-10] = 0
+            print(list(zip(*np.nonzero(diff))))
+            print(diff[np.nonzero(diff)])
+            np.testing.assert_array_almost_equal(
+                mp2_initial_point._t2, pyscf_mp.t2, decimal=10
+            )
 
 
 if __name__ == "__main__":
